requirements.txt
PySide6
opencv-python
PyYAML
